requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
gTTS==2.0.4
gTTS-token==1.1.3
idna==2.8
pyaspeller==0.1.0
PyAudio==0.2.11
pygame==1.9.6
pyqt5-tools==5.5.1.1.2
PySide==1.2.4
python-vlc==3.0.7110
requests==2.21.0
six==1.12.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
urllib3==1.24.3
